requirements.txt
astroid==3.3.5
dill==0.3.9
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
mypy==1.13.0
mypy-extensions==1.0.0
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
pylint==3.3.1
pytest==8.3.3
remotetypes @ file:///home/marcos/Escritorio/Universidad/Dist/venv/ssdd-remote-types
ruff==0.8.0
tomlkit==0.13.2
typing_extensions==4.12.2
zeroc-ice==3.7.10.1
kafka-python==2.0.2
confluent-kafka==2.2.0

